+++ tests/test_items.py
@@ -199,6 +199,7 @@ def test_array_multiline():
     assert expected == t.as_string()
 
     t = item([])
+    t.multiline(True)
 
     assert "[]" == t.as_string()