+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [0.8.0] - 2021-12-20
 
 ### Changed
 
@@ -221,7 +221,8 @@
 - Fixed handling of super tables with different sections.
 - Fixed raw strings escaping.
 
-[unreleased]: https://github.com/sdispater/tomlkit/compare/0.7.2...master
+[unreleased]: https://github.com/sdispater/tomlkit/compare/0.8.0...master
+[0.8.0]: https://github.com/sdispater/tomlkit/releases/tag/0.8.0
 [0.7.2]: https://github.com/sdispater/tomlkit/releases/tag/0.7.2
 [0.7.1]: https://github.com/sdispater/tomlkit/releases/tag/0.7.1
 [0.7.0]: https://github.com/sdispater/tomlkit/releases/tag/0.7.0